+++ src/main/java/com/jazeee/ddp/listeners/IDDPHeartbeatListener.java
@@ -1,4 +1,4 @@
-package com.jazeee.ddp;
+package com.jazeee.ddp.listeners;
 
 public interface IDDPHeartbeatListener {
