+++ src/main/java/com/jazeee/ddp/listeners/IDDPMethodListener.java
@@ -1,4 +1,4 @@
-package com.jazeee.ddp;
+package com.jazeee.ddp.listeners;
 
 import java.util.Map;
